+++ eve_neo4j/tests/test_settings_neo4j.py
@@ -36,7 +36,8 @@ people = {
             'type': 'integer'
         },
         'firstname': {
-            'type': 'string'
+            'type': 'string',
+            'unique': True
         },
         'lastname': {
             'type': 'string'